Removes signal handler for signal 28 (terminal resize signal)
The cleanup handler in this application was reponding to any and all process signals.
The intent was for cleanup to run regardless of how the application was terminated
(SIGTERM, SIGKILL, etc.).  This behavior is inappropriate because certain signals, like
28, are simply used to notify the process of system events.  Signal 28 (SIGWINCH) is
used to signal a terminal resize event.

It is likely the console blanking feature on the Pi is also sending this signal, which
would cause a undesired shutdown of the program.